Update the team counts based on user input
This allows the user to change the number of characters that have been
selected and the number of players in the game and see the numbers
update.

This also includes an update so that private properties are now all
protected - I think they're better.

Added a simple cache lookup to get one element from another, since I use
that functionality a lot.