Move the data to folders by locale
The idea is that the gulp task can loop over all the folders and take
all the contents, creating a "data" file for each locale that contains
all the information that the system will need to create the localised
content.

There will need to be a "meta" file as well that contains things like
the name of the locale, that way the gulp task can create a list of all
supported locales that the app can then use.