Allow info tokens to editted
This also shows the process for updating data:

 - The Store has the data.

 - The Model gets a copy of it.

 - The Controller reads the Model's data, passing it to the view.

 - The Model doesn't update its own data. Instead, it updates the Store
   and listens for the Store changing - as the Store changes its data,
   the Model updates its own, possibly updating the Controller to update
   the View

By doing this, the Store is the single source of truth and can be used
to save information in something like localStorage to remember
information between resets. It also allows any other Model to update its
own data because of the Store update.