Add custom script section to the scripts
Eventually the forms for official scripts and custom scripts will be in
their own tabs.

This commit also includes the addition of an "i18n" section in each
locale which will be used to store any text that JavaScript will need to
know. At the moment in includes a translation for "Custom Script", which
will eventually be dropped as it can be added to the HTML.