Update the Night Order rendering
The rendering now adds a placeholder HTML comment for all Night Order
entries, swapping it out for a viewable element if it should be visible.
This allows the element to be added and removed from the DOM without
having to re-draw the entire list, only modifying the elements that are
affected, so the drawing is far more efficient.