Start converting Movable into an observer
This is mainly so that the data is handled correctly. What should happen
is that Movable updates the Model, the Model updates the Store. The
update to the Store triggers an event that the Model responds to,
triggering its own event. The Controller hears the Model's event and
passes the data on to the View, which updates the element.

Although convoluted, the process would mean that the Store knows about
the position of seats (and reminders) so that it can always be set again
if the page is refreshed.