Combine the observers
The observers are now part of a central observer. This (theoretically)
allows the observers to be replaced with something else in the future
(in practice, I think I need to make the observer-like methods a mixin
first) and allows for a future (hypothetical) debugging tool to listen
for all events that are triggered.